requirements.txt
requests==2.31.0
python-dotenv==0.19.0
